a working k-unbound prototype

+++ include/struct.h
@@ -29,7 +29,7 @@ struct UserInput { // a container for user input
     
     std::vector<std::string> kmerDB; // a database of kmers (or DBG)
     std::string prefix = ".", outFile = "";
-    uint8_t kmerLen = 21;
+    uint32_t kmerLen = 21;
     
 };
